+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ocapld ChangeLog
 
+### Changed
+- Update webpack and babel.
+
 ## 1.1.0 - 2019-03-27
 
 ### Changed